fix(suite-native): ultimate solution to the passphrase pin, allow paassphser on device + pin on T1B1
